requirements.txt
torch>=2.0.0
torchvision>=0.15.0
monai>=1.3.0
monai-generative>=0.2.0
diffusers>=0.21.0
matplotlib>=3.6.0
tqdm>=4.64.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
tensorboard>=2.10.0
wandb>=0.13.0
hydra-core>=1.3.0
omegaconf>=2.3.0